perf: add user layout
